Add files via upload
In this project i used Cross-validation as a statistical method to evaluate generalization performance in a more stable and thorough way than using a split into training and test set. In cross-validation, instead of splitting the data set in to a training set and a test set, the data is split repeatedly and multiple models are trained. The most commonly used version of cross-validation is k-fold cross-validation, where k is a user specified number, usually five or ten. When performing five-fold crossvalidation, the data is first partitioned into five parts of (approximately) equal size, called folds.